--- assistant_launch/launch/.placeholder
@@ -1,1 +1,0 @@
-
