+++ assistant_navigation/maps/hector_suc6.yaml
@@ -1,0 +1,7 @@
+image: /home/user/workspace/turtlebot3/base/src/assistant/assistant_navigation/maps/hector_suc6.pgm
+resolution: 0.025000
+origin: [-25.612499, -25.612499, 0.000000]
+negate: 0
+occupied_thresh: 0.65
+free_thresh: 0.196
+
